bugfix: freeing a not previously alloced address during freeaddrinfo
The previous design used a single memory allocation to hold
both address and sin structs. freeaddrinfo on OSX doesn't expect
these to be inside the same physical allocation, and generates
a fatal error during the later call to freeaddrinfo(), that a
free() was calloc for a non-previously-allocated region.

Split the single allocation into two allocations.